requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
bcrypt==4.3.0
certifi==2025.4.26
click==8.2.0
dj-database-url==2.3.0
Django==5.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pytest==8.3.5
python-dotenv==1.1.0
python-jose==3.4.0
rsa==4.9.1
ruff==0.11.10
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
